tests in right place and working. Double check that the self urls are not stored
